fix: Address pre-commit failures
